+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Add `validateHistogram` and `debugPrint`
 - Fix a pointy centroid bug.
 - Add `Data.TDigest.NonEmpty` module
+- Add `mean`, `variance`, `stddev`